agrego ejercicio a usuario y se almacena en calendario

+++ src/main/java/com/tallerwebi/dominio/RepositorioEjercicio.java
@@ -5,7 +5,6 @@ import java.util.List;
 public interface RepositorioEjercicio {
 
     List<Ejercicio> obtenerTodosLosEjercicios();
-    Ejercicio obtenerEjercicioPorId(Integer id);
-
+    Ejercicio obtenerEjercicioPorId(Long id);
 
 }
